Feedback: - kdump_reader.c only compiled when respective flag is enabled - drgn_program_set_kdump() the only thing exposed through kdump_header.h - has_kdump_signature() moved to program.c - dump OSRELEASE_LEN macro - kdump_read() impossible short read without error - no need to clear buffer before using kdump_read() - has_kdump_signature() return NULL for cores smaller than KDUMP_SIG_LEN
