+++ libdrgn/kdump_reader.h
@@ -19,10 +19,8 @@
 #ifndef DRGN_KDUMP_READER_H
 #define DRGN_KDUMP_READER_H
 
-#ifdef LIBKDUMPFILE
 #include <stdbool.h>
 #include <stdint.h>
-#include <libkdumpfile/kdumpfile.h>
 
 #include "internal.h"
 
@@ -31,48 +29,35 @@
  *
  * @defgroup KdumpReader kdump reader
  *
- * kdump reader implementation of the MemoryReader interface.
+ * Logic for setting up drgn for reading kdump crash dumps.
  *
  * @{
  */
 
 #define	KDUMP_SIGNATURE	"KDUMP   "
-#define	SIG_LEN	(sizeof (KDUMP_SIGNATURE) - 1)
+#define	KDUMP_SIG_LEN	(sizeof (KDUMP_SIGNATURE) - 1)
 
-/** Check contents from the file descriptor start with the KDUMP signature */
-struct drgn_error *has_kdump_signature(int fd, bool *ret);
+#ifndef LIBKDUMPFILE
 
-/** Create a kdump-context from the given file descriptor */
-struct drgn_error *drgn_kdump_init(kdump_ctx_t **ctx, int fd);
+static inline struct drgn_error *
+drgn_program_set_kdump(const struct drgn_program *prog)
+{
+        return drgn_error_create(DRGN_ERROR_INVALID_ARGUMENT,
+	                         "drgn configured without libkdumpfile");
+}
 
-/** Clean up all kdump-related metadata. */
-void drgn_kdump_close(kdump_ctx_t *ctx);
-
-/**
- * Get all the vmcoreinfo from the kdump in a single string.
- *
- * @param[in] ctx Kdump context.
- * @param[out] ret vmcore info.
- * @return @c NULL on success, non-@c NULL on error.
- */
-struct drgn_error *drgn_kdump_get_raw_vmcoreinfo(kdump_ctx_t *ctx,
-                                                 const char **ret);
+#else /* LIBKDUMPFILE */
 
 /**
- * Get the expected architecture flags from the kdump file.
+ * Setup program-related context leveraging libkdumpfile so
+ * we are able to read from makedumpfile program dumps.
  *
- * @param[in] ctx Kdump context.
- * @param[out] arch architecture flags.
  * @return @c NULL on success, non-@c NULL on error.
  */
-struct drgn_error *drgn_kdump_get_arch(kdump_ctx_t *ctx,
-                                       enum drgn_architecture_flags *arch);
+struct drgn_error *drgn_program_set_kdump(struct drgn_program *prog);
 
-/** @ref drgn_memory_read_fn which reads from a kdump file. */
-struct drgn_error *drgn_read_kdump(void *buf, uint64_t address, size_t count,
-                                   uint64_t offset, void *arg, bool physical);
+#endif /* LIBKDUMPFILE */
 
 /** @} */
-#endif /* LIBKDUMPFILE */
 
 #endif /* DRGN_KDUMP_READER_H */
